requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-ai>=0.1.0
google-genai>=0.3.0
python-dotenv>=1.0.0
gradio>=4.0.0
requests>=2.31.0
filetype>=1.2.0
arize-phoenix>=12.2.0
openinference-instrumentation-pydantic-ai>=0.1.7
pytest>=8.4.2
pytest-asyncio>=1.2.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
uv>=0.7.8
pydantic-evals>=1.1.0
tenacity>=8.0.0